+++ src/Serializer/Normalizer/AbstractNormalizer.php
@@ -2,6 +2,7 @@
 
 namespace App\Serializer\Normalizer;
 
+use App\Service\Solr;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
@@ -12,7 +13,8 @@ class AbstractNormalizer
     public function __construct(
         #[Autowire(service: 'serializer.normalizer.object')]
         protected NormalizerInterface $decorated,
-        protected EntityManagerInterface $entityManager
+        protected EntityManagerInterface $entityManager,
+        protected Solr $solrService
 
     ) {
     }